An attempt to store each block in the chain in a separate file
This feature will help with scaling issue of the blockchain.
Current algorithm loads a chain from a singlle JSON file into a single
python dictionary.
This feature of separating blocks into many files is to decrease RAM
consumption.
However, by the time this commit is created, it is impossible to mine
because of FileNotFoundError in mine() function.